Refactor Rust API for speech denoiser
