+++ sherpa-onnx/rust/sherpa-onnx/src/lib.rs
@@ -7,7 +7,6 @@ mod online_asr;
 mod online_punctuation;
 mod online_speech_denoiser;
 mod speaker_embedding;
-mod speech_denoiser;
 mod tts;
 mod utils;
 mod vad;
@@ -22,7 +21,6 @@ pub use online_asr::*;
 pub use online_punctuation::*;
 pub use online_speech_denoiser::*;
 pub use speaker_embedding::*;
-pub use speech_denoiser::*;
 pub use tts::*;
 pub use utils::*;
 pub use vad::*;
